move mock data generation into separate class
